+++ src/page/room/services/demo.js
@@ -1,15 +1,17 @@
 import Service from '@home/core/baseClass/Service';
-import { http } from '@home/util/http';
+import injectHttp from '@home/core/inject/injectHttp';
 
 import Launcher from '@home/core/launcher';
 
-export default class DemoService extends Service {
+export default
+@injectHttp('roomApp')
+class DemoService extends Service {
     @Launcher.store('roomApp') store;
 
     @Launcher.store() stores;
 
     getInfo() {
-        return http.send({
+        return this.$http.send({
             url: 'https://api.github.com/users/123',
             method: 'get',
         });